chefbot/requirements.txt
flake8==7.1.2
openai==1.63.2
python-dotenv==1.0.1 
slack-bolt==1.22.0
slackstyler==0.0.3
